requirements.txt
psycopg2
social-auth-app-django

python-oauth2